+++ features/fixtures/mazerunner/app/src/main/java/com/bugsnag/android/mazerunner/MainActivity.kt
@@ -6,11 +6,15 @@ import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Build
 import android.os.Bundle
+import android.view.Window
 import android.widget.Button
 import android.widget.EditText
-import com.bugsnag.android.Configuration
 import com.bugsnag.android.mazerunner.scenarios.Scenario
+import org.json.JSONObject
 import java.io.File
+import java.lang.IllegalArgumentException
+import java.net.URL
+import kotlin.concurrent.thread
 
 class MainActivity : Activity() {
 
@@ -21,6 +25,7 @@ class MainActivity : Activity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        requestWindowFeature(Window.FEATURE_NO_TITLE)
         setContentView(R.layout.activity_main)
         log("Launched mazerunner fixture MainActivity")
         prefs = getPreferences(Context.MODE_PRIVATE)
@@ -29,41 +34,6 @@ class MainActivity : Activity() {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
             val closeDialog = Intent(Intent.ACTION_CLOSE_SYSTEM_DIALOGS)
             sendBroadcast(closeDialog)
-        }
-
-        // Clear persistent data (used to stop scenarios bleeding into each other)
-        findViewById<Button>(R.id.clear_persistent_data).setOnClickListener {
-            clearFolder("last-run-info")
-            clearFolder("bugsnag-errors")
-            clearFolder("device-id")
-            clearFolder("user-info")
-            clearFolder("fake")
-        }
-
-        // load the scenario first, which initialises bugsnag without running any crashy code
-        findViewById<Button>(R.id.start_bugsnag).setOnClickListener {
-            scenario = loadScenarioFromUi()
-            scenario?.startBugsnag(true)
-        }
-
-        // execute the pre-loaded scenario, or load it then execute it if needed
-        findViewById<Button>(R.id.run_scenario).setOnClickListener {
-            if (scenario == null) {
-                scenario = loadScenarioFromUi()
-                scenario?.startBugsnag(false)
-            }
-
-            /**
-             * Enqueues the test case with a delay on the main thread. This avoids the Activity wrapping
-             * unhandled Exceptions
-             */
-            window.decorView.postDelayed(
-                {
-                    log("Executing scenario")
-                    scenario?.startScenario()
-                },
-                1
-            )
         }
 
         val clearUserData = findViewById<Button>(R.id.clearUserData)
@@ -81,6 +51,95 @@ class MainActivity : Activity() {
             apiKeyField.text.clear()
             apiKeyField.text.append(apiKey)
         }
+
+        startCommandRunner()
+    }
+
+    // Starts a thread to poll for Maze Runner actions to perform
+    private fun startCommandRunner() {
+        // Get the next maze runner command
+        thread(start = true) {
+            while (true) {
+                Thread.sleep(1000)
+                try {
+                    // Get the next command from Maze Runner
+                    val commandUrl: String = "http://bs-local.com:9339/command"
+                    val commandStr = URL(commandUrl).readText()
+                    if (commandStr == "null") {
+                        log("No Maze Runner commands queued")
+                        continue
+                    }
+
+                    // Log the received command
+                    log("Received command: $commandStr")
+                    var command = JSONObject(commandStr)
+                    val action = command.getString("action")
+                    val scenarioName = command.getString("scenario_name")
+                    val scenarioMode = command.getString("scenario_mode")
+                    val sessionsUrl = command.getString("sessions_endpoint")
+                    val notifyUrl = command.getString("notify_endpoint")
+                    log("command.action: $action")
+                    log("command.scenarioName: $scenarioName")
+                    log("command.scenarioMode: $scenarioMode")
+                    log("command.sessionsUrl: $sessionsUrl")
+                    log("command.notifyUrl: $notifyUrl")
+
+                    runOnUiThread {
+                        // Display some feedback of the action being run on he UI
+                        val actionField = findViewById<EditText>(R.id.command_action)
+                        val scenarioField = findViewById<EditText>(R.id.command_scenario)
+                        actionField.setText(action)
+                        scenarioField.setText(scenarioName)
+
+                        // Perform the given action on the UI thread
+                        when (action) {
+                            "start_bugsnag" -> startBugsnag(scenarioName, scenarioMode, sessionsUrl, notifyUrl)
+                            "run_scenario" -> runScenario(scenarioName, scenarioMode, sessionsUrl, notifyUrl)
+                            "clear_persistent_data" -> clearPersistentData()
+                            else -> throw IllegalArgumentException("Unknown action: $action")
+                        }
+                    }
+                } catch (e: Exception) {
+                    log("Failed to fetch command from Maze Runner", e)
+                }
+            }
+        }
+    }
+
+    // load the scenario first, which initialises bugsnag without running any crashy code
+    private fun startBugsnag(eventType: String, mode: String, sessionsUrl: String, notifyUrl: String) {
+        scenario = loadScenario(eventType, mode, sessionsUrl, notifyUrl)
+        scenario?.startBugsnag(true)
+    }
+
+    // execute the pre-loaded scenario, or load it then execute it if needed
+    private fun runScenario(eventType: String, mode: String, sessionsUrl: String, notifyUrl: String) {
+        if (scenario == null) {
+            scenario = loadScenario(eventType, mode, sessionsUrl, notifyUrl)
+            scenario?.startBugsnag(false)
+        }
+
+        /**
+         * Enqueues the test case with a delay on the main thread. This avoids the Activity wrapping
+         * unhandled Exceptions
+         */
+        window.decorView.postDelayed(
+            {
+                log("Executing scenario")
+                scenario?.startScenario()
+            },
+            1
+        )
+    }
+
+    // Clear persistent data (used to stop scenarios bleeding into each other)
+    private fun clearPersistentData() {
+        log("Clearing persistent data")
+        clearFolder("last-run-info")
+        clearFolder("bugsnag-errors")
+        clearFolder("device-id")
+        clearFolder("user-info")
+        clearFolder("fake")
     }
 
     private fun clearFolder(name: String) {
@@ -90,20 +149,9 @@ class MainActivity : Activity() {
         folder.deleteRecursively()
     }
 
-    private fun loadScenarioFromUi(): Scenario {
-        val scenarioPicker = findViewById<EditText>(R.id.scenario_name)
-        val eventType = scenarioPicker.text.toString()
-        val eventMetadata = findViewById<EditText>(R.id.scenario_metadata)
-        val metadata = eventMetadata.text.toString()
+    private fun loadScenario(eventType: String, mode: String, sessionsUrl: String, notifyUrl: String): Scenario {
 
-        val config = loadConfigFromUi()
-        return Scenario.load(this, config, eventType, metadata)
-    }
-
-    private fun loadConfigFromUi(): Configuration {
         val apiKeyField = findViewById<EditText>(R.id.manualApiKey)
-        val notifyEndpointField = findViewById<EditText>(R.id.notify_endpoint)
-        val sessionEndpointField = findViewById<EditText>(R.id.session_endpoint)
 
         val manualMode = apiKeyField.text.isNotEmpty()
         val apiKey = when {
@@ -115,12 +163,11 @@ class MainActivity : Activity() {
             log("Running in manual mode with API key: $apiKey")
             setStoredApiKey(apiKey)
         }
-        val notify = notifyEndpointField.text.toString()
-        val sessions = sessionEndpointField.text.toString()
-        return prepareConfig(apiKey, notify, sessions) {
+        val config = prepareConfig(apiKey, notifyUrl, sessionsUrl) {
             val interceptedLogMessages = scenario?.getInterceptedLogMessages()
             interceptedLogMessages?.contains(it) ?: false
         }
+        return Scenario.load(this, config, eventType, mode)
     }
 
     private fun apiKeyStored() = prefs.contains(apiKeyKey)