+++ features/support/env.rb
@@ -4,6 +4,12 @@ BeforeAll do
   Maze.config.receive_requests_wait = 60
 end
 
+Before do
+  $scenario_mode = ''
+  $sessions_endpoint = 'http://bs-local.com:9339/sessions'
+  $notify_endpoint = 'http://bs-local.com:9339/notify'
+end
+
 Before('@skip') do |scenario|
   skip_this_scenario("Skipping scenario")
 end